bring back link assignment
