+++ src/lapis/entities.rs
@@ -1,5 +1,4 @@
-use crate::{interaction::*, lapis::*};
-use bevy::sprite::AlphaMode2d;
+use crate::lapis::*;
 
 //TODO joints not spawned programmatically are not accessible
 //TODO do we need Vec<Entity>? can we spawn structures like grids?
@@ -157,280 +156,4 @@ fn method_entity(expr: &ExprMethodCall, lapis: &mut Lapis) -> Option<Entity> {
         _ => return None,
     }
     Some(e)
-}
-
-// ---- observers ----
-
-#[derive(Event, Clone)]
-pub enum Property {
-    X(f32),
-    Y(f32),
-    Rx(f32),
-    Ry(f32),
-    Rot(f32),
-    Mass(f32),
-    Vx(f32),
-    Vy(f32),
-    Va(f32),
-    Restitution(f32),
-    LinDamp(f32),
-    AngDamp(f32),
-    Inertia(f32),
-    H(f32),
-    S(f32),
-    L(f32),
-    A(f32),
-    Sides(u32),
-    Cmx(f32),
-    Cmy(f32),
-    Friction(f32),
-    Tail(usize),
-    Layer(u32),
-    Dynamic(bool),
-    Sensor(bool),
-    Links(String),
-    CodeI(String),
-    CodeF(String),
-}
-
-pub fn set_observer(
-    trig: Trigger<Property>,
-    mut trans_query: Query<&mut Transform, With<RigidBody>>,
-    mut commands: Commands,
-    mut lin_velocity_query: Query<&mut LinearVelocity>,
-    mesh_ids: Query<&Mesh2d>,
-    mut meshes: ResMut<Assets<Mesh>>,
-    material_ids: Query<&MeshMaterial2d<ColorMaterial>>,
-    mut materials: ResMut<Assets<ColorMaterial>>,
-    mut cm_query: Query<&mut CenterOfMass>,
-    mut tail_query: Query<&mut Tail>,
-    mut code_query: Query<&mut Code>,
-    selected_query: Query<Entity, With<Selected>>,
-) {
-    let e = trig.entity();
-    // methods applied to PLACEHOLDER affect the selected entities
-    if e == Entity::PLACEHOLDER {
-        let mut targets = Vec::new();
-        for e in selected_query.iter() {
-            targets.push(e);
-        }
-        commands.trigger_targets(trig.event().clone(), targets);
-        return;
-    }
-    match *trig.event() {
-        Property::X(val) => {
-            if let Ok(mut t) = trans_query.get_mut(e) {
-                t.translation.x = val;
-            }
-        }
-        Property::Y(val) => {
-            if let Ok(mut t) = trans_query.get_mut(e) {
-                t.translation.y = val;
-            }
-        }
-        Property::Rx(val) => {
-            if let Ok(mut t) = trans_query.get_mut(e) {
-                t.scale.x = val;
-            }
-        }
-        Property::Ry(val) => {
-            if let Ok(mut t) = trans_query.get_mut(e) {
-                t.scale.y = val;
-            }
-        }
-        Property::Rot(val) => {
-            if let Ok(mut t) = trans_query.get_mut(e) {
-                t.rotation = Quat::from_rotation_z(val);
-            }
-        }
-        Property::Mass(val) => {
-            if trans_query.contains(e) {
-                commands.entity(e).insert(Mass(val));
-            }
-        }
-        Property::Vx(val) => {
-            if let Ok(mut v) = lin_velocity_query.get_mut(e) {
-                v.x = val;
-            }
-        }
-        Property::Vy(val) => {
-            if let Ok(mut v) = lin_velocity_query.get_mut(e) {
-                v.y = val;
-            }
-        }
-        Property::Va(val) => {
-            if trans_query.contains(e) {
-                commands.entity(e).insert(AngularVelocity(val));
-            }
-        }
-        Property::Restitution(val) => {
-            if trans_query.contains(e) {
-                commands.entity(e).insert(Restitution::new(val));
-            }
-        }
-        Property::LinDamp(val) => {
-            if trans_query.contains(e) {
-                commands.entity(e).insert(LinearDamping(val));
-            }
-        }
-        Property::AngDamp(val) => {
-            if trans_query.contains(e) {
-                commands.entity(e).insert(AngularDamping(val));
-            }
-        }
-        Property::Inertia(val) => {
-            if trans_query.contains(e) {
-                commands.entity(e).insert(AngularInertia(val));
-            }
-        }
-        Property::H(val) => {
-            if let Ok(mat_id) = material_ids.get(e) {
-                let mat = materials.get_mut(mat_id).unwrap();
-                let mut hsla: Hsla = mat.color.into();
-                hsla.hue = val;
-                mat.color = hsla.into();
-            }
-        }
-        Property::S(val) => {
-            if let Ok(mat_id) = material_ids.get(e) {
-                let mat = materials.get_mut(mat_id).unwrap();
-                let mut hsla: Hsla = mat.color.into();
-                hsla.saturation = val;
-                mat.color = hsla.into();
-            }
-        }
-        Property::L(val) => {
-            if let Ok(mat_id) = material_ids.get(e) {
-                let mat = materials.get_mut(mat_id).unwrap();
-                let mut hsla: Hsla = mat.color.into();
-                hsla.lightness = val;
-                mat.color = hsla.into();
-            }
-        }
-        Property::A(val) => {
-            if let Ok(mat_id) = material_ids.get(e) {
-                let mat = materials.get_mut(mat_id).unwrap();
-                let mut hsla: Hsla = mat.color.into();
-                hsla.alpha = val;
-                mat.color = hsla.into();
-            }
-        }
-        Property::Sides(val) => {
-            let val = val.clamp(3, 512);
-            if let Ok(mesh_id) = mesh_ids.get(e) {
-                let mesh = meshes.get_mut(mesh_id).unwrap();
-                *mesh = RegularPolygon::new(1., val).into();
-                commands
-                    .entity(e)
-                    .insert((Sides(val), Collider::regular_polygon(1., val)));
-            }
-        }
-        Property::Cmx(val) => {
-            if let Ok(mut v) = cm_query.get_mut(e) {
-                v.x = val;
-            }
-        }
-        Property::Cmy(val) => {
-            if let Ok(mut v) = cm_query.get_mut(e) {
-                v.y = val;
-            }
-        }
-        Property::Friction(val) => {
-            if trans_query.contains(e) {
-                commands.entity(e).insert(Friction::new(val));
-            }
-        }
-        Property::Tail(val) => {
-            if let Ok(mut tail) = tail_query.get_mut(e) {
-                tail.len = val;
-            }
-        }
-        Property::Layer(val) => {
-            if trans_query.contains(e) {
-                let layer = 1 << val;
-                commands
-                    .entity(e)
-                    .insert(CollisionLayers::from_bits(layer, layer));
-            }
-        }
-        Property::Dynamic(val) => {
-            if trans_query.contains(e) {
-                if val {
-                    commands.entity(e).insert(RigidBody::Dynamic);
-                } else {
-                    commands.entity(e).insert(RigidBody::Static);
-                }
-            }
-        }
-        Property::Sensor(val) => {
-            if trans_query.contains(e) {
-                if val {
-                    commands.entity(e).insert(Sensor);
-                } else {
-                    commands.entity(e).remove::<Sensor>();
-                }
-            }
-        }
-        Property::Links(ref val) => {
-            if trans_query.contains(e) {
-                commands.entity(e).insert(Links(val.clone()));
-            }
-        }
-        Property::CodeI(ref val) => {
-            if let Ok(mut c) = code_query.get_mut(e) {
-                c.0 = val.clone();
-            }
-        }
-        Property::CodeF(ref val) => {
-            if let Ok(mut c) = code_query.get_mut(e) {
-                c.1 = val.clone();
-            }
-        }
-    }
-}
-
-#[derive(Event)]
-pub struct InsertDefaults(f32);
-
-pub fn insert_defaults(
-    trig: Trigger<InsertDefaults>,
-    mut commands: Commands,
-    mut meshes: ResMut<Assets<Mesh>>,
-    mut materials: ResMut<Assets<ColorMaterial>>,
-    settings: Res<DrawSettings>,
-) {
-    let e = trig.entity();
-    let r = trig.event().0;
-    let material = ColorMaterial {
-        color: Srgba::from_u8_array(settings.color).into(),
-        alpha_mode: AlphaMode2d::Blend,
-        ..default()
-    };
-    let mesh_handle = meshes.add(RegularPolygon::new(1., settings.sides));
-    let mat_handle = materials.add(material);
-    let layer = 1 << settings.collision_layer;
-    commands.entity(e).insert((
-        Mesh2d(mesh_handle),
-        MeshMaterial2d(mat_handle),
-        settings.rigid_body,
-        Links(settings.links.clone()),
-        Code(settings.code.0.clone(), settings.code.1.clone()),
-        Mass(r * r * r),
-        AngularInertia(r * r * r),
-        CenterOfMass(settings.center_of_mass),
-        Collider::regular_polygon(1., settings.sides),
-        CollisionLayers::from_bits(layer, layer),
-        (
-            LinearDamping(settings.lin_damp),
-            AngularDamping(settings.ang_damp),
-            Restitution::new(settings.restitution),
-            Friction::new(settings.friction),
-        ),
-        Transform::from_scale(Vec3::new(r, r, 1.)),
-        Sides(settings.sides),
-        Tail {
-            len: settings.tail,
-            ..default()
-        },
-    ));
 }
